Fix #1
The plugin has been modified to check the slug name against the slugs
present in wp_terms. If the slug name is already present an error will
be displayed and the status will not be saved.

If any existing statuses have a core WC slug, then it will have to be
manually deleted.