Fix #3
The plugin inherits the WC_Settings_Page class. When the settings are
saved, the wc_clean() function is called which uses the
sanitize_text_field() to sanitize the data. Since we do not have any
control over the WC functions, it makes sense to change the merge tags
to {customer_email} & {admin_email}.

The sanitize_text() removes %ad from %admin% when saving the settings.
Hence we've now changed the merge tags from %customer% to {custom_email}
& %admin% to {admin_email}

For existing users when the plugin is updated the merge tags will be
updated in the options table.

Even though emails can't be sent in the Lite version, the settings can
be created and saved. Hence we need to ensure the changes are reflected
correctly in the UI as well as the DB.

Also, the update code has been included to ensure existing data is
updated for this version as well and the code continues to be in sync
with PRO.