Fix #4
Added 2 new merge tags for the email content. {first_name} and
{last_name}

The feature to send email is not present in Lite. However, the UI is
available and settings can be saved. So the same needs to be tested.

This code has been added to ensure the 2 versions continue to be in
sync.